backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
requests
python-multipart
fastapi[all]